+++ R/negate.R
@@ -15,11 +15,5 @@
 #' # Same as
 #' x %>% discard("y") %>% length()
 negate <- function(.p, .default = FALSE) {
-  .p <- as_mapper(.p)
-
-  body(.p) <- expr({
-    !(!!body(.p))
-  })
-
-  .p
+  compose(`!`, as_mapper(.p))
 }