+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # purrr 0.3.0.9000
 
+* `negate()` now works with generic functions and functions with early
+  returns.
+
 * `compose()` now works with generic functions again (#629, #639). Its
   set of unit tests was expanded to cover many edge cases.